Moving logic into Trait
